requirements.txt
requests==2.20.0
Flask==0.12.2
flask-restplus==0.12.1
Flask_JWT_Extended==3.7.2
Flask_Bcrypt==0.7.1
python-dotenv
elasticsearch
google-cloud-language